refactor(call): apply conventions
