--- src/app/call/components/controls.module.css
@@ -1,25 +1,0 @@
-/*
-controls.module.css
-Ojos Project
-*/
-.controls {
-  display: flex;
-  gap: 32px;
-  justify-content: space-evenly;
-  width: 100%;
-  max-width: 900px;
-  position: fixed;
-  bottom: 5%;
-  /* https://stackoverflow.com/questions/3157372/css-horizontal-centering-of-a-fixed-div */
-  left: 50%;
-  transform: translateX(-50%);
-}
-
-.video_container {
-  height: 90dvh;
-  border: solid 2px black;
-  border-radius: 20px;
-
-  min-width: 500px;
-  min-height: 500px;
-}
