+++ src/app/call/page.tsx
@@ -1,10 +1,14 @@
-// call/page.tsx
-// Ojos Project
+/**
+ * File:     call/page.tsx
+ * Purpose:  The main calling page. Wraps around the WebcamRecorder component.
+ * Authors:  Ojos Project & Iris contributors
+ * License:  GNU General Public License v3.0
+ */
 "use client";
-import classes from "./page.module.css";
-import Button from "../../components/Button";
+import styles from "./page.module.css";
+import Button from "@/components/Button";
 import { invoke } from "@tauri-apps/api/core";
-import WebcamRecorder from "./components/controls";
+import WebcamRecorder from "./_components/WebcamRecorder";
 import useKeyPress from "@/components/useKeyPress";
 import { useRouter } from "next/navigation";
 
@@ -17,7 +21,7 @@ export default function Call() {
 
   return (
     <>
-      <header className={classes.header}>
+      <header className={styles.header}>
         <Button
           type="SECONDARY"
           label="View Recordings"
@@ -26,7 +30,7 @@ export default function Call() {
           }}
         />
       </header>
-      <main className={classes.video_main}>
+      <main className={styles.videoMain}>
         <WebcamRecorder />
       </main>
     </>
